Make FileReader.py framework API and restructure entire library
FileReader.py now has a framework of the future API. Unfortunately, it
wasn't able to be run without putting the entire project under the
folder "PythonFileLibrary/", which is why this introduces two
major changes.

In addition to this, .gitignore was optimized to include **/__pycache__
to make sure the folder didn't appear anywhere in this project.